requirements.txt
graphics.py==5.0.1.post1
mpmath==1.0.0
sympy==1.3
